Possible solution for exec return status issue
This attempts to solve the problem by bubbling up the exit status form init process to intermediate
and from intermediate to the main/ exec calling process. The init process's image will be replaced after
exec call, and the intermediate process will wait for this process to exit. Once exited, the intemediate
process itself will exit with the same exit status as the init process. THe exec calling process
(the main process) will wait for this intermediate process, and will exit with the its status.

This also allows calling multiple execs for the same container, without us needing to generate random hash
For each pipe, which will potentially return the exit status.